feat: template for events
