+++ frontend/src/common/enums/app-path.ts
@@ -1,0 +1,7 @@
+const appPath = {
+	ROOT: "/",
+	MY_EVENTS: "/my_events",
+	CREATE_EVENT: "/create-event",
+};
+
+export { appPath };
