--- frontend/src/common/enums/appPath.ts
@@ -1,8 +1,0 @@
-
-const appPath = {
-    ROOT: "/",
-    MY_EVENTS: "/my_events",
-    CREATE_EVENT: "/create-event"
-}
-
-export { appPath };
