--- frontend/src/pages/HomePage/components/EventItem/EventItem.tsx
@@ -1,13 +1,0 @@
-import type React from "react";
-import type { Event } from "../../../../common/types";
-
-const EventItem: React.FC<Event> = ({ title, description, date, location, visibility, capacity }) => {
-
-    return (
-        <>
-            
-        </>
-    )
-}
-
-export { EventItem };
