+++ package.json
@@ -4,7 +4,8 @@
 	"main": "index.js",
 	"lint-staged": {
 		"**/*.{ts,tsx}": [
-			"biome lint --write"		]
+			"biome lint --write"
+		]
 	},
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
